Implement variableLevels for the generic join extenders
Every generic-join extender now reports the levels its variables occupy
(as Longs, so Clojure level indices pass through unchanged):

- GenericOrPrefixExtender keys its per-branch tries by variable levels
  instead of participation levels, fixing branch leaking when branch
  decisions depend on variables bound outside the OR; the totalLevels
  constructor parameter is now derived from the children
- GenericNotPrefixExtender takes the not pattern's variable levels and
  participates at the last of them instead of the query's last level
- enable the or/and branch-leak test case

+++ src/main/kotlin/org/hooray/iterator/AVLPrefixExtender.kt
@@ -13,7 +13,7 @@ sealed interface AVLIndex {
 }
 
 @Suppress("UNCHECKED_CAST")
-class AVLPrefixExtender(val avlIndex: AVLIndex, participatesInLevel: List<Int>) : GenericPrefixExtender(
+class AVLPrefixExtender(val avlIndex: AVLIndex, participatesInLevel: List<Long>) : GenericPrefixExtender(
     when(avlIndex) {
         is AVLIndex.AVLMapIndex -> SealedIndex.MapIndex(avlIndex.map as Map<Any, Any>)
         is AVLIndex.AVLSetIndex -> SealedIndex.SetIndex(avlIndex.set as Set<Any>)
